Refactor: 커밋 데이터 불러오기 및 전처리 코드 분리
